refactor(dialog): split DialogView into Form renderer + modal shell
Extract the declarative section/action renderer out of DialogView into a
new Form.svelte; DialogView becomes the modal shell (backdrop + centered
card) that wraps <Form>. The future creation panel will be a second surface
wrapping the same Form.

Pure renderer refactor: no behavior change, no shared-type change, no IPC
change. DialogHost and src/shared/dialog-types.ts are untouched. Section/
action/event tests move to Form.test.ts; DialogView.test.ts is slimmed to
surface-only assertions.
